+++ Assets/InfiniteScrollView/Scripts/Runtime/HorizontalInfiniteScrollView.cs
@@ -1,5 +1,4 @@
 ﻿using Cysharp.Threading.Tasks;
-using System;
 using UnityEngine;
 
 namespace HowTungTung
@@ -7,23 +6,23 @@ namespace HowTungTung
     public class HorizontalInfiniteScrollView : InfiniteScrollView
     {
         public float spacing;
-        public bool isAtLeft = true;
-        public bool isAtRight = true;
-
-        public override async UniTask InitializePool(object args = null)
-        {
-            await base.InitializePool(args);
-            isAtLeft = true;
-            isAtRight = true;
-        }
 
         public override void RefreshCellVisibility()
         {
             if (dataList.Count == 0)
                 return;
+
+            // Viewport
             float viewportInterval = scrollRect.viewport.rect.width;
-            float minViewport = -scrollRect.content.anchoredPosition.x;
+
+            // Check content direction pivot
+            if (this._contentDirCoeff == 0) this._contentDirCoeff = scrollRect.content.pivot.x > 0 ? 1f : -1f;
+
+            // Set content direction
+            float minViewport = scrollRect.content.anchoredPosition.x * this._contentDirCoeff;
             Vector2 viewportRange = new Vector2(minViewport - extendVisibleRange, minViewport + viewportInterval + extendVisibleRange);
+
+            // Hide
             float contentWidth = padding.left;
             for (int i = 0; i < dataList.Count; i++)
             {
@@ -34,29 +33,44 @@ namespace HowTungTung
                 }
                 contentWidth += dataList[i].cellSize.x + spacing;
             }
+
+            // Show
             contentWidth = padding.left;
             for (int i = 0; i < dataList.Count; i++)
             {
                 var visibleRange = new Vector2(contentWidth, contentWidth + dataList[i].cellSize.x);
                 if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
                 {
-                    SetupCell(i, new Vector2(contentWidth, -(padding.top - padding.bottom)));
+                    InfiniteCell cell = null;
+                    if (cellList[i] == null)
+                    {
+                        if (_cellPool.Count > 0) cell = _cellPool.Dequeue();
+                        else Debug.Log("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
+                    }
+                    // Check cell direciton pivot
+                    float dirCoeff = 1f;
+                    if (cell != null) dirCoeff = cell.RectTransform.pivot.x > 0 ? -1f : 1f;
+                    SetupCell(cell, i, new Vector2(contentWidth * dirCoeff, -(padding.top - padding.bottom)));
                     if (visibleRange.y >= viewportRange.x)
-                        cellList[i].transform.SetAsLastSibling();
+                        cellList[i]?.transform.SetAsLastSibling();
                     else
-                        cellList[i].transform.SetAsFirstSibling();
+                        cellList[i]?.transform.SetAsFirstSibling();
                 }
                 contentWidth += dataList[i].cellSize.x + spacing;
             }
+
+            // Check scroll position
             if (scrollRect.content.sizeDelta.x > viewportInterval)
             {
-                isAtLeft = viewportRange.x + extendVisibleRange + dataList[0].cellSize.x <= dataList[0].cellSize.x;
-                isAtRight = scrollRect.content.sizeDelta.x - viewportRange.y + extendVisibleRange + dataList[dataList.Count - 1].cellSize.x <= dataList[dataList.Count - 1].cellSize.x;
+                this._isAtLeft = viewportRange.x + extendVisibleRange + dataList[0].cellSize.x <= dataList[0].cellSize.x;
+                this._isAtRight = scrollRect.content.sizeDelta.x - viewportRange.y + extendVisibleRange + dataList[dataList.Count - 1].cellSize.x <= dataList[dataList.Count - 1].cellSize.x;
             }
             else
             {
-                isAtLeft = true;
-                isAtRight = true;
+                this._isAtTop = false;
+                this._isAtBottom = false;
+                this._isAtLeft = true;
+                this._isAtRight = true;
             }
         }
 
@@ -120,7 +134,8 @@ namespace HowTungTung
 
             if (scrollRect.content.anchoredPosition.x != width)
             {
-                DoSnapping(new Vector2(-width, 0), duration);
+                // Check content direction pivot
+                DoSnapping(new Vector2(width * this._contentDirCoeff, 0), duration);
             }
         }