+++ Assets/InfiniteScrollView/Scripts/Runtime/VerticalInfiniteScrollView.cs
@@ -1,5 +1,4 @@
 ﻿using Cysharp.Threading.Tasks;
-using System;
 using UnityEngine;
 
 namespace HowTungTung
@@ -7,25 +6,23 @@ namespace HowTungTung
     public class VerticalInfiniteScrollView : InfiniteScrollView
     {
         public float spacing;
-        public bool isAtTop = true;
-        public bool isAtBottom = true;
-
-        public override async UniTask InitializePool(object args = null)
-        {
-            if (this.IsInitialized) return;
-
-            await base.InitializePool(args);
-            isAtTop = true;
-            isAtBottom = true;
-        }
 
         public override void RefreshCellVisibility()
         {
             if (dataList.Count == 0)
                 return;
+
+            // Viewport
             float viewportInterval = scrollRect.viewport.rect.height;
-            float minViewport = scrollRect.content.anchoredPosition.y;
+
+            // Check content direction pivot
+            if (this._contentDirCoeff == 0) this._contentDirCoeff = scrollRect.content.pivot.y > 0 ? 1f : -1f;
+
+            // Set content direction
+            float minViewport = scrollRect.content.anchoredPosition.y * this._contentDirCoeff;
             Vector2 viewportRange = new Vector2(minViewport - extendVisibleRange, minViewport + viewportInterval + extendVisibleRange);
+
+            // Hide
             float contentHeight = padding.top;
             for (int i = 0; i < dataList.Count; i++)
             {
@@ -36,29 +33,44 @@ namespace HowTungTung
                 }
                 contentHeight += dataList[i].cellSize.y + spacing;
             }
+
+            // Show
             contentHeight = padding.top;
             for (int i = 0; i < dataList.Count; i++)
             {
                 var visibleRange = new Vector2(contentHeight, contentHeight + dataList[i].cellSize.y);
                 if (visibleRange.y >= viewportRange.x && visibleRange.x <= viewportRange.y)
                 {
-                    SetupCell(i, new Vector2(padding.left - padding.right, -contentHeight));
+                    InfiniteCell cell = null;
+                    if (cellList[i] == null)
+                    {
+                        if (_cellPool.Count > 0) cell = _cellPool.Dequeue();
+                        else Debug.Log("<color=#ff4242>The cell display error occurred, not enough cells in the cell pool!!!</color>");
+                    }
+                    // Check cell direciton pivot
+                    float dirCoeff = 1f;
+                    if (cell != null) dirCoeff = cell.RectTransform.pivot.y > 0 ? -1f : 1f;
+                    SetupCell(cell, i, new Vector2(padding.left - padding.right, contentHeight * dirCoeff));
                     if (visibleRange.y >= viewportRange.x)
-                        cellList[i].transform.SetAsLastSibling();
+                        cellList[i]?.transform.SetAsLastSibling();
                     else
-                        cellList[i].transform.SetAsFirstSibling();
+                        cellList[i]?.transform.SetAsFirstSibling();
                 }
                 contentHeight += dataList[i].cellSize.y + spacing;
             }
+
+            // Check scroll position
             if (scrollRect.content.sizeDelta.y > viewportInterval)
             {
-                isAtTop = viewportRange.x + extendVisibleRange <= 0.001f;
-                isAtBottom = scrollRect.content.sizeDelta.y - viewportRange.y + extendVisibleRange <= 0.001f;
+                this._isAtTop = viewportRange.x + extendVisibleRange <= 0.001f;
+                this._isAtBottom = scrollRect.content.sizeDelta.y - viewportRange.y + extendVisibleRange <= 0.001f;
             }
             else
             {
-                isAtTop = true;
-                isAtBottom = true;
+                this._isAtTop = true;
+                this._isAtBottom = true;
+                this._isAtLeft = false;
+                this._isAtRight = false;
             }
         }
 
@@ -122,7 +134,8 @@ namespace HowTungTung
 
             if (scrollRect.content.anchoredPosition.y != height)
             {
-                DoSnapping(new Vector2(0, height), duration);
+                // Check content direction pivot
+                DoSnapping(new Vector2(0, height * this._contentDirCoeff), duration);
             }
         }