+++ Assets/InfiniteScrollView/Scripts/Samples~/01_Vertical/TestGUI_01.cs
@@ -1,7 +1,6 @@
 ﻿using Cysharp.Threading.Tasks;
 using HowTungTung;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 public class TestGUI_01 : MonoBehaviour
 {
@@ -19,10 +18,7 @@ public class TestGUI_01 : MonoBehaviour
 
     private void OnGUI()
     {
-        if (GUILayout.Button("NextScene"))
-        {
-            SceneManager.LoadScene((int)Mathf.Repeat(SceneManager.GetActiveScene().buildIndex + 1, SceneManager.sceneCountInBuildSettings));
-        }
+        #region Add
         if (GUILayout.Button("Add 100 Random Height Cell"))
         {
             for (int i = 0; i < 100; i++)
@@ -32,6 +28,7 @@ public class TestGUI_01 : MonoBehaviour
             }
             infiniteScrollView.Refresh();
         }
+
         if (GUILayout.Button("Add"))
         {
             var data = new InfiniteCellData(new Vector2(0, 50));
@@ -39,17 +36,37 @@ public class TestGUI_01 : MonoBehaviour
             infiniteScrollView.Refresh();
             infiniteScrollView.SnapLast(0.1f);
         }
+        #endregion
+
+        #region Remove
         GUILayout.Label("Remove Index");
         removeIndex = GUILayout.TextField(removeIndex);
         if (GUILayout.Button("Remove"))
         {
             infiniteScrollView.Remove(int.Parse(removeIndex)).Forget();
         }
+        #endregion
+
+        #region Snap
         GUILayout.Label("Snap Index");
         snapIndex = GUILayout.TextField(snapIndex);
         if (GUILayout.Button("Snap"))
         {
             infiniteScrollView.Snap(int.Parse(snapIndex), 0.1f);
         }
+        #endregion
+
+        #region Scroll
+        GUILayout.Label("Vertical");
+        if (GUILayout.Button("Scroll to top"))
+        {
+            infiniteScrollView.ScrollToTop();
+        }
+
+        if (GUILayout.Button("Scroll to bottom"))
+        {
+            infiniteScrollView.ScrollToBottom();
+        }
+        #endregion
     }
 }