+++ Assets/InfiniteScrollView/Scripts/Samples~/02_Horizontal/TestGUI_02.cs
@@ -1,7 +1,6 @@
 ﻿using Cysharp.Threading.Tasks;
 using HowTungTung;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 public class TestGUI_02 : MonoBehaviour
 {
@@ -19,10 +18,7 @@ public class TestGUI_02 : MonoBehaviour
 
     private void OnGUI()
     {
-        if (GUILayout.Button("NextScene"))
-        {
-            SceneManager.LoadScene((int)Mathf.Repeat(SceneManager.GetActiveScene().buildIndex + 1, SceneManager.sceneCountInBuildSettings));
-        }
+        #region Add
         if (GUILayout.Button("Add 100 Random Width Cell"))
         {
             for (int i = 0; i < 100; i++)
@@ -31,6 +27,7 @@ public class TestGUI_02 : MonoBehaviour
             }
             infiniteScrollView.Refresh();
         }
+
         GUILayout.Label("Add New Cell Width");
         if (GUILayout.Button("Add"))
         {
@@ -38,17 +35,37 @@ public class TestGUI_02 : MonoBehaviour
             infiniteScrollView.Refresh();
             infiniteScrollView.SnapLast(0.1f);
         }
+        #endregion
+
+        #region Remove
         GUILayout.Label("Remove Index");
         removeIndex = GUILayout.TextField(removeIndex);
         if (GUILayout.Button("Remove"))
         {
             infiniteScrollView.Remove(int.Parse(removeIndex)).Forget();
         }
+        #endregion
+
+        #region Snap
         GUILayout.Label("Snap Index");
         snapIndex = GUILayout.TextField(snapIndex);
         if (GUILayout.Button("Snap"))
         {
             infiniteScrollView.Snap(int.Parse(snapIndex), 0.1f);
         }
+        #endregion
+
+        #region Scroll
+        GUILayout.Label("Horizontal");
+        if (GUILayout.Button("Scroll to left"))
+        {
+            infiniteScrollView.ScrollToLeft();
+        }
+
+        if (GUILayout.Button("Scroll to right"))
+        {
+            infiniteScrollView.ScrollToRight();
+        }
+        #endregion
     }
 }