+++ Assets/InfiniteScrollView/Scripts/Samples~/03_Grid/TestGUI_03.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
+﻿using Cysharp.Threading.Tasks;
 using HowTungTung;
-using Cysharp.Threading.Tasks;
+using UnityEngine;
 
 public class TestGUI_03 : MonoBehaviour
 {
@@ -28,10 +24,7 @@ public class TestGUI_03 : MonoBehaviour
 
     private void OnGUI()
     {
-        if (GUILayout.Button("NextScene"))
-        {
-            SceneManager.LoadScene((int)Mathf.Repeat(SceneManager.GetActiveScene().buildIndex + 1, SceneManager.sceneCountInBuildSettings));
-        }
+        #region Add
         if (GUILayout.Button("Add 1000 Cell"))
         {
             for (int i = 0; i < 1000; i++)
@@ -40,6 +33,7 @@ public class TestGUI_03 : MonoBehaviour
             }
             infiniteScrollView.Refresh();
         }
+
         if (GUILayout.Button("Add"))
         {
             var data = new InfiniteCellData(new Vector2(100, 100));
@@ -47,17 +41,48 @@ public class TestGUI_03 : MonoBehaviour
             infiniteScrollView.Refresh();
             infiniteScrollView.SnapLast(0.1f);
         }
+        #endregion
+
+        #region Remove
         GUILayout.Label("Remove Index");
         removeIndex = GUILayout.TextField(removeIndex);
         if (GUILayout.Button("Remove"))
         {
             infiniteScrollView.Remove(int.Parse(removeIndex)).Forget();
         }
+        #endregion
+
+        #region Snap
         GUILayout.Label("Snap Index");
         snapIndex = GUILayout.TextField(snapIndex);
         if (GUILayout.Button("Snap"))
         {
             infiniteScrollView.Snap(int.Parse(snapIndex), 0.1f);
         }
+        #endregion
+
+        #region Scroll
+        GUILayout.Label("Vertical");
+        if (GUILayout.Button("Scroll to top"))
+        {
+            infiniteScrollView.ScrollToTop();
+        }
+
+        if (GUILayout.Button("Scroll to bottom"))
+        {
+            infiniteScrollView.ScrollToBottom();
+        }
+
+        GUILayout.Label("Horizontal");
+        if (GUILayout.Button("Scroll to left"))
+        {
+            infiniteScrollView.ScrollToLeft();
+        }
+
+        if (GUILayout.Button("Scroll to right"))
+        {
+            infiniteScrollView.ScrollToRight();
+        }
+        #endregion
     }
 }